Core feature done. Needs gimmicks.
The core function of the preview is done. Needs a sidebar for choosing models from /public/models.

The Canvas also needs a dynamic window sizing function ( a loop ) to update the viewport whenever client sizes their browser.

Also need to integrate a Backend service so I am thinking of Firebase unless we have a better solution.